feat(breakout-rooms): implement auto-setup functionality
- Add new feature for automatic breakout room setup
- Implement auto-assign action and room creation logic
- Add support for removing all rooms and adding new ones
- Integrate auto-setup with existing breakout room functionality
- Update UI to show auto-setup button when appropriate
