+++ react/features/app/middlewares.web.ts
@@ -11,6 +11,7 @@ import '../no-audio-signal/middleware';
 import '../notifications/middleware';
 import '../noise-detection/middleware';
 import '../breakout-room-presetup/middleware';
+import '../breakout-room-autosetup/middleware';
 import '../old-client-notification/middleware';
 import '../power-monitor/middleware';
 import '../prejoin/middleware';
