+++ react/features/app/types.ts
@@ -28,6 +28,7 @@ import { ITestingState } from '../base/testing/reducer';
 import { ITooltipState } from '../base/tooltip/reducer';
 import { INoSrcDataState, ITracksState } from '../base/tracks/reducer';
 import { IUserInteractionState } from '../base/user-interaction/reducer';
+import { IAutosetupBreakoutRoomsState } from '../breakout-room-autosetup/reducer';
 import { IPresetBreakoutRoomsState } from '../breakout-room-presetup/reducer';
 import { IBreakoutRoomsState } from '../breakout-rooms/reducer';
 import { ICalendarSyncState } from '../calendar-sync/reducer';
@@ -119,6 +120,7 @@ export interface IReduxState {
     'features/base/tooltip': ITooltipState;
     'features/base/tracks': ITracksState;
     'features/base/user-interaction': IUserInteractionState;
+    'features/breakout-room-autosetup': IAutosetupBreakoutRoomsState;
     'features/breakout-room-presetup': IPresetBreakoutRoomsState;
     'features/breakout-rooms': IBreakoutRoomsState;
     'features/calendar-sync': ICalendarSyncState;
