+++ react/features/base/participants/types.ts
@@ -21,14 +21,6 @@ export interface IParticipant {
     fakeParticipant?: FakeParticipant;
     features?: IParticipantFeatures;
     getId?: Function;
-    /**
-     * Add this field to store the iDigest UserId, Related modifications:
-     *  1. For Local:
-     *      1.1 _user2participant：react/features/base/jwt/middleware.ts:323
-     *      1.2 _overwriteLocalParticipant：react/features/base/jwt/middleware.ts:105
-     *  2. For Remote: _participantJoined：react/features/base/participants/reducer.ts:645.
-     */
-    iDigestId?: string;
     id: string;
     isJigasi?: boolean;
     isPromoted?: boolean;
