+++ react/features/breakout-room-autosetup/actionTypes.ts
@@ -1,0 +1,1 @@
+export const _AVAILABLE_AUTO_SET_BREAKOUT_ROOMS = '_AVAILABLE_AUTO_SET_BREAKOUT_ROOMS';
