+++ react/features/breakout-room-autosetup/components/AutoBreakoutRoomButton.tsx
