+++ react/features/breakout-room-autosetup/constants.ts
@@ -1,0 +1,4 @@
+/**
+ * Key for this feature.
+ */
+export const FEATURE_KEY = 'features/breakout-room-autosetup';
