+++ react/features/breakout-room-presetup/actionTypes.ts
@@ -9,5 +9,3 @@ export const _PRESET_BREAKOUT_ROOMS_ADD_LISTENER = '_PRESET_BREAKOUT_ROOMS_ADD_L
 export const _PRESET_BREAKOUT_ROOMS_CLEAN_LISTENER = '_PRESET_BREAKOUT_ROOMS_CLEAN_LISTENER';
 
 export const _AVAILABLE_TO_SET_BREAKOUT_ROOMS = '_AVAILABLE_TO_SET_BREAKOUT_ROOMS';
-
-export const _AVAILABLE_AUTO_SET_BREAKOUT_ROOMS = '_AVAILABLE_AUTO_SET_BREAKOUT_ROOMS';
