+++ react/features/participants-pane/components/web/ParticipantsPane.tsx
@@ -13,9 +13,10 @@ import Button from '../../../base/ui/components/web/Button';
 import ClickableIcon from '../../../base/ui/components/web/ClickableIcon';
 import { BUTTON_TYPES } from '../../../base/ui/constants.web';
 import { findAncestorByClass } from '../../../base/ui/functions.web';
-import AutoBreakoutRoomButton from '../../../breakout-room-presetup/components/AutoBreakoutRoomButton';
+import AutoBreakoutRoomButton from '../../../breakout-room-autosetup/components/AutoBreakoutRoomButton';
+import { isAutoBreakoutRoomButtonVisible } from '../../../breakout-room-autosetup/functions';
 import LoadPresetBreakoutRoomButton from '../../../breakout-room-presetup/components/LoadPresetBreakoutRoomButton';
-import { isAutoBreakoutRoomButtonVisible, isPresetBreakoutRoomButtonVisible } from '../../../breakout-room-presetup/functions';
+import { isPresetBreakoutRoomButtonVisible } from '../../../breakout-room-presetup/functions';
 import { isAddBreakoutRoomButtonVisible } from '../../../breakout-rooms/functions';
 import MuteEveryoneDialog from '../../../video-menu/components/web/MuteEveryoneDialog';
 import { shouldDisplayCurrentVisitorsList } from '../../../visitors/functions';
