+++ tsconfig.native.json
@@ -36,6 +36,7 @@
         "react/features/web-hid",
         "react/features/whiteboard",
         "react/features/breakout-room-presetup",
+        "react/features/breakout-room-autosetup",
         "**/web/*",
         "**/*.web.ts",
         "**/*.web.tsx"
